requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
easydict==1.12
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
faiss-gpu==1.7.2
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
h11==0.14.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.15.0
onnxruntime-gpu==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pretrainedmodels==0.7.4
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
segmentation-models-pytorch==0.3.3
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
timm==0.9.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==1.3.1
torchvision==0.13.1+cu113
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
zipp==3.17.0
